reorganize, remove most of useless code
