+++ lua/nvim-treesitter/health.lua
@@ -1,9 +1,9 @@
 local api = vim.api
 local fn = vim.fn
 
-local queries = require'nvim-treesitter.query'
 local locals = require'nvim-treesitter.locals'
 local configs = require'nvim-treesitter.configs'
+local queries = require'nvim-treesitter.query'
 
 local health_start = vim.fn["health#report_start"]
 local health_ok = vim.fn['health#report_ok']
@@ -12,6 +12,16 @@ local health_warn = vim.fn['health#report_warn']
 local health_error = vim.fn['health#report_error']
 
 local M = {}
+
+local function queries_health(lang)
+  if not queries.get_query(lang, "locals") then
+    health_warn("No `locals.scm` query found for " .. lang, {
+      "Open an issue at https://github.com/nvim-treesitter/nvim-treesitter"
+    })
+  else
+    health_ok("`locals.scm` found.")
+  end
+end
 
 local function install_health()
   if fn.executable('git') == 0 then
@@ -42,19 +52,19 @@ function M.checkhealth()
 
   local missing_parsers = {}
   -- Parser installation checks
-  for _, parser_name in pairs(configs.available_parsers()) do
-    local installed = #api.nvim_get_runtime_file('parser/'..parser_name..'.so', false)
+  for _, ft in pairs(configs.available_parsers()) do
+    local installed = #api.nvim_get_runtime_file('parser/'..ft..'.so', false)
 
     -- Only print informations about installed parsers
     if installed == 1 then
-      health_start(parser_name .. " parser healthcheck")
-      health_ok(parser_name .. " parser found.")
+      health_start(ft .. " parser healthcheck")
+      health_ok(ft .. " parser found.")
 
-      locals.checkhealth(parser_name)
+      queries_health(ft)
     elseif installed > 1 then
-      health_warn(string.format("Multiple parsers found for %s, only %s will be used.", parser_name, installed[1]))
+      health_warn(string.format("Multiple parsers found for %s, only %s will be used.", ft, installed[1]))
     else
-      table.insert(missing_parsers, parser_name)
+      table.insert(missing_parsers, ft)
     end
   end
