+++ lua/nvim-treesitter/locals.lua
@@ -4,27 +4,11 @@
 local api = vim.api
 local ts = vim.treesitter
 local queries = require'nvim-treesitter.query'
-local parsers = require'nvim-treesitter.parsers'
+local utils = require'nvim-treesitter.utils'
 
 local M = {
   locals={}
 }
-
-function M.checkhealth(lang)
-  local health_start = vim.fn["health#report_start"]
-  local health_ok = vim.fn['health#report_ok']
-  local health_info = vim.fn['health#report_info']
-  local health_warn = vim.fn['health#report_warn']
-  local health_error = vim.fn['health#report_error']
-
-  if not queries.get_query(lang, "locals") then
-    health_warn("No `locals.scm` query found for " .. lang, {
-      "Open an issue at https://github.com/nvim-treesitter/nvim-treesitter"
-    })
-  else
-    health_ok("`locals.scm` found.")
-  end
-end
 
 function M.collect_locals(bufnr)
   local ft = api.nvim_buf_get_option(bufnr, "ft")
@@ -33,7 +17,7 @@ function M.collect_locals(bufnr)
   local query = queries.get_query(ft, 'locals')
   if not query then return end
 
-  local parser = parsers.get_parser(bufnr, ft)
+  local parser = utils.get_parser(bufnr, ft)
   if not parser then return end
 
   local root = parser:parse():root()
