+++ plugin/nvim-treesitter.vim
@@ -5,16 +5,13 @@ if exists('g:loaded_nvim_treesitter')
 endif
 
 augroup NvimTreesitter
-augroup END
+augroup end
 
 let g:loaded_nvim_treesitter = 1
 
 lua << EOF
 ts_installable_parsers = function()
   return table.concat(require'nvim-treesitter.configs'.available_parsers(), '\n')
-end
-ts_available_modules = function()
-  return table.concat(require'nvim-treesitter.configs'.available_modules(), '\n')
 end
 require'nvim-treesitter'.setup()
 EOF
